Cleanup + Debt Handling

+++ src/virtual/StarCamera.cpp
@@ -6,37 +6,37 @@
 #define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/transform.hpp>
 
-star::StarCamera::StarCamera(const float& width, const float& height) : resolution(glm::vec2{width, height}){
-
+star::StarCamera::StarCamera(const uint32_t &width, const uint32_t &height) : resolution(glm::ivec2{width, height})
+{
 }
 
-glm::mat4 star::StarCamera::getViewMatrix() const{
-    glm::vec3 position = this->getPosition(); 
+glm::mat4 star::StarCamera::getViewMatrix() const
+{
+    glm::vec3 position = this->getPosition();
 
-    return glm::lookAt(
-        this->getPosition(),
-        this->getPosition() + glm::vec3(this->getForwardVector()),
-        glm::vec3(this->getUpVector())
-    );
+    return glm::lookAt(this->getPosition(), this->getPosition() + glm::vec3(this->getForwardVector()),
+                       glm::vec3(this->getUpVector()));
 }
 
 glm::mat4 star::StarCamera::getProjectionMatrix() const{
-    return glm::perspective(getVerticalFieldOfView(true), this->resolution.x / this->resolution.y, this->nearClippingPlaneDistance, this->farClippingPlaneDistance);
+    return glm::perspective(getVerticalFieldOfView(true), (float)this->resolution.x / (float)this->resolution.y, this->nearClippingPlaneDistance, this->farClippingPlaneDistance);
 }
 
-float star::StarCamera::getHorizontalFieldOfView(const bool& inRadians) const{
+float star::StarCamera::getHorizontalFieldOfView(const bool &inRadians) const
+{
     if (inRadians)
         return glm::radians(this->horizontalFieldOfView);
     else
-        return this->horizontalFieldOfView; 
+        return this->horizontalFieldOfView;
 }
 
-float star::StarCamera::getVerticalFieldOfView(const bool& inRadians) const{
+float star::StarCamera::getVerticalFieldOfView(const bool &inRadians) const
+{
     const float verticalRadians = glm::radians(this->horizontalFieldOfView);
-    const float resultRadians = 2.0f * glm::atan(glm::tan(verticalRadians / 2.0f) / (this->resolution.x / this->resolution.y));
+    const float resultRadians = 2.0f * glm::atan(glm::tan(verticalRadians / 2.0f) / (float)(this->resolution.x / (float)this->resolution.y));
 
     if (inRadians)
-        return resultRadians; 
+        return resultRadians;
     else
-        return glm::degrees(resultRadians); 
+        return glm::degrees(resultRadians);
 }
